backend/requirements.txt
fastapi
uvicorn
beanie
motor
pydantic
python-jose
passlib[bcrypt]
python-dotenv
